requirements.txt
asgiref==3.2.10
Django==3.0.7
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
django-tinymce==3.0.2
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
